requirements.txt
requests
python-dotenv
oracledb
markdown
weasyprint
flask
